* changed length of content in packets of Stealth<->External Script Communication + added parsing of cliloc arguments in "tooltip" gump element, available in internal and external scripts

+++ ScriptSDK/Stealth API/Network/Packet.cs
@@ -9,7 +9,7 @@ namespace StealthAPI
     internal class Packet
     {
         internal byte[] Data;
-        internal short DataLength;
+        internal int DataLength;
         internal PacketType Method;
 
         public Packet()
